requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docx==0.2.4
environs==14.1.1
exceptiongroup==1.2.2
fastapi==0.115.8
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
lxml==5.3.1
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.1.0
openai==1.63.0
packaging==24.2
pillow==11.1.0
polib==1.2.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-gettext==5.0
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
yarl==1.18.3
